requirements.txt
ipykernel
pandas
Faker
tabulate
weaviate-client
langchain-core
langgraph>0.2.27
langchain-openai
langchain-community